example/requirements.txt
git+https://github.com/adaptivdesign/django-sellmo.git

Pillow
sorl-thumbnail
django-widget-tweaks
django-grappelli
